split intnt filter and middleware
