+++ aiogram_dialog/api/entities/middleware.py
@@ -1,0 +1,4 @@
+STORAGE_KEY = "aiogd_storage_proxy"
+STACK_KEY = "aiogd_stack"
+CONTEXT_KEY = "aiogd_context"
+CALLBACK_DATA_KEY = "aiogd_original_callback_data"
